requirements.txt
certifi==2019.11.28
Click==7.1.1
cmdstanpy==0.4.0
convertdate==2.1.3
cycler==0.10.0
Cython==0.29.15
cryptography==2.8
dill==0.3.0
ephem==3.7.7.0
fastapi==0.61.1
flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
h11==0.9.0
holidays==0.10.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
lunardate==0.2.0
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.2
patsy==0.5.1
prometheus-client==0.7.1
pydantic==1.6.1
pyjwt==1.7.1
pyparsing==2.4.6
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
setuptools-git==1.2
six==1.14.0
starlette==0.13.6
statsmodels==0.11.1
starlette-prometheus==0.7.0
tornado==6.0.4
tqdm==4.43.0
uvicorn==0.11.8
waitress==1.4.3
websockets==8.1
Werkzeug==1.0.0
pytest==5.4.1
pytest-cov==2.8.1
requests==2.23.0
psutil==5.7.2